pyproject.toml
[project]
name = "dormitorydigitaltwin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "influxdb3-python>=0.17.0",
    "matplotlib>=3.10.8",
    "neo4j>=5.0.0",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    # Chat interface dependencies
    "google-generativeai>=0.8.0",
    "langfuse>=2.0.0",
    "streamlit>=1.40.0",
]
